dispatcher: keep the GDBusConnection instance in the global variable
It's anyway a singleton that is still referenced by other components.
So unrefing it in the mainloop does not actually release any memory.

However, the GDBusConnection singleton is fundamental for the run of
the program. Keep it accessible in the global variables.

Note that soon I will drop the GDBusInterfaceSkeleton and only operate
on the GDBusConnection. Then it makes more sense to keep it around.

Note that fundamentally we want to keep the amount of global state
small. But this connection is anyway a singleton (that we already
implicitly use). So, it doesn't change the amount of global state
nor does it really have much state (we either have the singleton or we
don't).